Benchmarks implemented
